requirements.txt
Flask==1.1.1
gunicorn==20.0.4
h5py==2.10.0
imutils==0.5.3
Jinja2==2.11.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.1
Pillow==7.0.0
requests==2.23.0
requests-oauthlib==1.2.0
scikit-learn==0.22.1
scipy==1.4.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
Theano==1.0.4


